move "path" function to more appropriate place
