+++ test/path.go
@@ -1,4 +1,4 @@
-package docker
+package test
 
 import (
 	"os"
